rename _colors.py to colors.py
